service

+++ src/orders/orders.service.ts
@@ -2,40 +2,63 @@ import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Order, orderDocument, orderSchema } from './orders.schema';
 import { Model } from 'mongoose';
+import { Client, ClientDocument } from 'src/clients/clients.schema';
 
 @Injectable()
 export class orderService {
-  constructor( @InjectModel(Order.name) private OrdersModel: Model<orderDocument>) {}
+  constructor(
+     @InjectModel(Order.name) private OrdersModel: Model<orderDocument>,
+    @InjectModel(Client.name) private ClientModel: Model<ClientDocument>
+) {}
 
-  async createorder(data: any): Promise<orderDocument> {
-    try {
-      const createdOrder = new this.OrdersModel(data);
-      return await createdOrder.save();
-    } catch (error) {
-      throw new HttpException(
-        { message: 'Failed to create order', error: error.message },
-        HttpStatus.INTERNAL_SERVER_ERROR,
+async findOrdersByFilters(filters: any, skip: number, limit: number): Promise<[orderDocument[], number]> {
+  const query: any = {};
+
+  if (filters.status) query.status = filters.status;
+  if (filters.paymentMethod) query.paymentMethod = filters.paymentMethod;
+  if (filters.clientId) query.clientId = filters.clientId;
+  if (filters.serviceId) query.serviceId = { $in: [filters.serviceId] };
+  if (filters.cartype) query.cartype = filters.cartype;
+
+  const [orders, total] = await Promise.all([
+    this.OrdersModel.find(query)
+      .populate('clientId')
+      .populate('serviceId')
+      .sort({ orderDate: -1 })
+      .skip(skip)
+      .limit(limit)
+      .exec(),
+
+    this.OrdersModel.countDocuments(query),
+  ]);
+
+  return [orders, total];
+}
+
+async createorder(data: any): Promise<orderDocument> {
+  try {
+    const createdOrder = new this.OrdersModel(data);
+    const savedOrder = await createdOrder.save();
+
+    if (data.clientId) {
+      await this.ClientModel.findByIdAndUpdate(
+        data.clientId,
+        { $push: { orders: savedOrder._id } },
+        { new: true }
       );
     }
-  }
 
-async findAllorder(): Promise<orderDocument[]> {
-  try {
-    return await this.OrdersModel.find({})
-      .populate(orderSchema.path('clientId'))       // جلب بيانات العميل
-      .populate('serviceId')      // جلب بيانات الخدمات
-      .sort({ orderDate: -1 })    
-      .exec();
+    return savedOrder;
+
   } catch (error) {
     throw new HttpException(
-      { message: 'Failed to fetch orders', error: error.message },
+      { message: 'Failed to create order', error: error.message },
       HttpStatus.INTERNAL_SERVER_ERROR,
     );
   }
 }
 
-
-  async findByIdorder(id: string): Promise<orderDocument> {
+  async findorderById(id: string): Promise<orderDocument> {
     try {
       const order = await this.OrdersModel.findById(id).exec();
       if (!order) {
@@ -69,8 +92,24 @@ async findAllorder(): Promise<orderDocument[]> {
     }
   }
 
+
   async updateorder(id: string, data: any): Promise<{ message: string }> {
     try {
+
+      const client= await this.ClientModel.findById(data.clientId).exec();
+       const order= await this.OrdersModel.findById(id).exec();
+      if (!client) {
+        throw new HttpException(
+          { message: 'Client not found' },
+          HttpStatus.NOT_FOUND,
+        );
+      }
+      if (order?.status === 'completed') {
+        throw new HttpException(
+          { message: 'sorry Order already completed' },
+          HttpStatus.NOT_FOUND,
+        );
+      }
       const updated = await this.OrdersModel.findByIdAndUpdate(id, data, {
         new: true,
       }).exec();
@@ -88,11 +127,78 @@ async findAllorder(): Promise<orderDocument[]> {
       );
     }
   }
+  
+async updateorderstatus(id: string, data: any): Promise<{ message: string }> {
+  try {
+    const client = await this.ClientModel.findById(data.clientId).exec();
+    const order = await this.OrdersModel.findById(id).exec();
 
-  async deleteorder(id: string): Promise<{ message: string }> {
+    if (!client) {
+      throw new HttpException(
+        { message: 'Client not found' },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    if (!order) {
+      throw new HttpException(
+        { message: 'Order not found' },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    // ✅ شرط السماح فقط لو اليوزر Admin
+    if (client.role !== 'admin') {
+      throw new HttpException(
+        { message: 'Only admin can update order status' },
+        HttpStatus.FORBIDDEN,
+      );
+    }
+    const updated = await this.OrdersModel.findByIdAndUpdate(id, data, {
+      new: true,
+    }).exec();
+
+    if (!updated) {
+      throw new HttpException(
+        { message: 'Order not found' },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    return { message: 'Order updated successfully' };
+
+  } catch (error) {
+    throw new HttpException(
+      { message: 'Failed to update order', error: error.message },
+      HttpStatus.INTERNAL_SERVER_ERROR,
+    );
+  }
+}
+
+  
+
+  async deleteorder(id: string,clientId:string): Promise<{ message: string }> {
     try {
+      const client= await this.ClientModel.findById(clientId).exec();
+       const order= await this.OrdersModel.findById(id).exec();
+      if (!client) {
+        throw new HttpException(
+          { message: 'Client not found' },
+          HttpStatus.NOT_FOUND,
+        );
+      }
+      if (order?.status === 'completed') {
+        throw new HttpException(
+          { message: 'sorry Order already completed You can not delete' },
+          HttpStatus.NOT_FOUND,
+        );
+      }
       const deleted = await this.OrdersModel.findByIdAndDelete(id).exec();
-      if (!deleted) {
+      const detetedfromClient = await this.ClientModel.findByIdAndUpdate(
+        clientId,
+        { $pull: { orders: id } },
+        { new: true }
+      )
+      if (!deleted||!detetedfromClient) {
         throw new HttpException(
           { message: 'Order not found' },
           HttpStatus.NOT_FOUND,
@@ -106,4 +212,5 @@ async findAllorder(): Promise<orderDocument[]> {
       );
     }
   }
+
 }
